Unification of basic types
This begins work on the core of a logic programming system. It has three
types of expression:

- variables, identified by name; a repeated variable in the same
  expression must unify to the same value everywhere it is used

- words: atomic values whose name is their value

- sequences: arbitrary-length sequences of expressions

All that exists so far is a `State` struct, which can unify two
expressions and return a new state with variables bound as appropriate.

There's a few places this could be improved to recurse over expressions
less often, and to clone less often, but we should have benchmarks
before changing that. We could also replace HashMap with the `im`
equivalent to reduce the cost of cloning a state.