Use `im::hashmap::HashMap` in place of `std::collections::HashMap`
Exploring the search space in a logic program generates a lot of
intermediate states, and these must be immutable, since a single state
is used as the basis for exploring many parallel search paths. We're
therefore cloning a `State` or the `HashMap` of its bindings on every
call to `State::unify`, before adding new variable bindings to the
clone.

It makes sense to use the `HashMap` structure from `im` to do this as
it's designed to make this pattern cheap -- it implements a purely
functional data structure that shares memory so that cloning and
mutation are less expensive.

On the benchmarks it appears to have made the very simple cases slower
but is making an improvement on the more complex cases. We expect this
to be net improvement over real programs where states become quite
large.

    test nested_seqs                    ... bench:      13,959 ns/iter (+/- 871)
    test nested_seqs_with_indirect_vars ... bench:      67,348 ns/iter (+/- 5,525)
    test seqs_with_vars                 ... bench:      12,386 ns/iter (+/- 706)
    test var                            ... bench:       2,362 ns/iter (+/- 204)
    test words                          ... bench:       1,142 ns/iter (+/- 142)