Wrap `Expr` values in `Rc` pointers
Unifying expressions involves a lot of cloning as modified copies of
expressions are built, especially when resolving a sequence to replace
its variables with their bound values. We can reduce the cost of this by
wrapping all expressions in an `Rc`, so that we only need to clone a
pointer, not the actual `Expr` structure.

This has made a dent on some of our benchmarks, notably reducing the
time for `nested_seqs_with_indirect_vars` and `seqs_with_vars` by about
1/3.

    test nested_seqs                    ... bench:       7,175 ns/iter (+/- 515)
    test nested_seqs_with_indirect_vars ... bench:      39,035 ns/iter (+/- 1,976)
    test seqs_with_vars                 ... bench:       6,144 ns/iter (+/- 368)
    test var                            ... bench:       1,516 ns/iter (+/- 112)
    test words                          ... bench:         508 ns/iter (+/- 238)