Resolve variables without building new expressions
While unifying expressions in `State::unify_mut` we currently call
`State::resolve` with each input `Expr`, which builds a copy of the
`Expr` with all bound variables recursively expanded to their values.
That is, if we call `resolve()` on the expression:

    a $x $y

in the `State` with bindings:

    $x => b $y
    $y => c

We will get:

    a (b c) c

Calling this recursive `resolve()` function from the recursive
`unify_mut()` means we're doing several recursions over the same
structure -- every sub-expression visited by `unify_mut()` triggers a
full recursive scan of itself with `resolve()`.

All we actually need in order to unify two expressions is to resolve any
variables we visit in `unify_mut()` to their current values. That is,
when looking at the sequence `a $x $y`, when we visit `$x` we resolve it
to `b $y`, and then continue walking. We don't need to expand `$y` in
place when we resolve `$x` because we'll resolve `$y` *when we visit
it*.

The only case for recursively expanding a variable is when it simply
points at another variable, that is in this state:

    $x => $y
    $y => c

Resolving `$x` should give us `c` not `$y`, so that we do not think `$y`
is unbound attempt to re-assign it to a new value.

`State::resolve` is reserved for the case where the caller wants a
completely resolved copy of an expression, and it is not needed for
unification.

This has resulting in a big improvement on some of our sequence
benchmarks...

Before:
    test nested_seqs                    ... bench:       7,223 ns/iter (+/- 434)
    test nested_seqs_with_indirect_vars ... bench:      30,914 ns/iter (+/- 1,557)
    test seqs_with_vars                 ... bench:       4,021 ns/iter (+/- 259)
    test var                            ... bench:       1,525 ns/iter (+/- 1,077)
    test words                          ... bench:         519 ns/iter (+/- 25)

After:
    test nested_seqs                    ... bench:         608 ns/iter (+/- 52)
    test nested_seqs_with_indirect_vars ... bench:       7,413 ns/iter (+/- 4,430)
    test seqs_with_vars                 ... bench:       3,112 ns/iter (+/- 213)
    test var                            ... bench:       1,528 ns/iter (+/- 111)
    test words                          ... bench:         506 ns/iter (+/- 33)