Interleave multiple streams
In a logic system, attempting to prove some statement can yield an
infinite number of proofs -- an infinite number of states where the
statement unifies with the existing rules. This poses problems for
proving things based on self-recursive rules.

For example, imagine we have two rules for subtyping: all types are
subtypes of themselves, and S is a subtype of U if S is a subtype of T,
and T is a subtype of U.

    Rule 1: $T <: $T
    Rule 2: if $S <: $T, and $T <: $U, then $S <: $U

Now, given these facts:

    Gerbil <: Rodent
    Rodent <: Mammal

Can we prove `Gerbil <: Mammal`?

First we try Rule 1, but `Gerbil <: Mammal` does not unify with `$T <:
$T`, so we try Rule 2.

Rule 2 gives us `Gerbil <: Mammal` if `Gerbil <: $T` and `$T <: Mammal`,
so we try to prove `Gerbil <: $T.

Trying Rule 1 works this time: `Gerbil <: $T` unifies with `$T <: $T` to
give `$T = Gerbil`. But Rule 2 also works: `Gerbil <: $U` gives us `$S =
Gerbil` and `$U = $T` with `$T` still unknown, so we recurse into Rule
2's first premise and try to derive `Gerbil <: $T`, but this is just
what we were trying to prove to start with.

This generates an infinite recursion and an infinite set of proofs for
`Gerbil <: $T` without telling us what `$T` is, so we get stuck
exploring this search path and never generate a useful result for
`Gerbil <: $T`.

A better strategy is to avoid a fully depth-first search and make it
partially breadth-first. We note that `Gerbil <: $T` unifies with three
of our system's rules:

- Rule 1, giving one state with `$T = Gerbil`
- Rule 2, giving an infinite set of states where `$T` is never bound
- The fact `Gerbil <: Rodent`, giving one state with `$T = Rodent`

Rather than getting stuck enumerating the states from Rule 2, we could
take the third result, `$T = Rodent`, and make progress with our
original goal. That is, we were originally trying to prove `Gerbil <:
$T` and `$T <: Mammal`, in order to prove `Gerbil <: Mammal`. If we're
able to discover the state where `$T = Rodent`, that gives us `Gerbil <:
Rodent` and `Rodent <: Mammal`, which are both stated facts and the
problem is solved.

We get stuck because using Rule 1 to prove `Gerbil <: $T` gives us
`Gerbil <: Gerbil` and `Gerbil <: Mammal` for Rule 2, and `Gerbil <:
Mammal` is what we were already trying to prove. Using Rule 2 similarly
gives us an infinite recursion. We need to skip past these two infinite
sinkholes and find `Gerbil <: Rodent` to make progress.

μKanren [1] does this by interleaving the streams of results formed by
matching a goal against each available rule. The streams might be
infinite, but if we can interleave them then we get some breadth-first
exploration of the search space and we're able to find proofs where a
purely depth-first search would fail.

We can do this by representing the result of an attempt to derive a
statement, as an infinite stream, using Rust's `Iterator` trait. This
makes it easy to generate lazy infinite streams of results whenever we
try to unify a statement with a particular rule. The `Interleave` struct
here takes an `Iterator` of `Iterator` and produces a flattened version
that takes a result from each inner `Iterator` in turn until all are
exhausted. For storage into a `Vec`, the outer `Iterator` must be
finite, but the inner ones may be infinite.

[1]: http://webyrd.net/scheme-2013/papers/HemannMuKanren2013.pdf