Sets of rules, just enough to write logic programs
So far we've focussed on getting `State::unify` working well, but to
write logic programs we need more than the ability to unify two given
expressions. We need to define sets of rules, so that any input
expression can be tried against all the available rules to generate
proofs.

The new `lang` module here defines `RuleSet`, which allows the
definition of rules with a name, a possibly-empty set of premises, and a
conclusion. Statements are proved by trying to unify them against all
existing conclusions, and for those that match, trying recursively to
prove the rule's premises.

To make this work, every invocation of a rule needs to generate a
distinct set of variables -- a variable `$x` in a rule must create a new
binding each time the rule is activated, rather than referring to the
same binding across activations. We add a simple scoping mechanism to
achieve this: in `Rule::match_target`, the conclusion and premises have
their variables replaced with copies suffixed with a monotonically
increasing number.

This "scope number" is the number of variable bindings in the current
`State`, which only increases and will be different across rule
invocations unless the rule's conclusion contains no variables -- such
rules are usually simple facts with no premises.

This copying of expressions re-introduces some performance issues we'd
previously removed from `State::unify` so we should look for ways to
improve how this scoping is done -- this is just a proof of concept to
get a system working. Some example rules and derivations are given in
`examples/append.rs`.