Flatten infinite streams of infinite streams
A few commits ago we introduced the `Interleave` type, which could
interleave a finite set of infinite streams. However in general we
actually need to be able to flatten *infinite sets of infinite streams*.

If a certain rule generates an infinite set of states in which it
matches the target expression, *each* of those states will in turn be
used as the basis to try to derive further expressions, and each of
those derivations may produce an infinite stream.

For example, imagine a rule has premises A and B and we try to derive A
in some state S, and this produces an infinite stream of states S'. We
will then try to derive B in *each state* in S', and each of *those*
derivations of B may themselves return an infinite stream. In order for
the rule as a whole to return `Iterator<Item=State>`, we need to flatten
these streams.

This is effectively Hilbert's Hotel [1]. We can solve it by naively
interleaving the first stream with the result of flattening all the
following streams, i.e.:

    flatten []     = []
    flatten (x:xs) = interleave x (flatten xs)

Or to put it another way:

    streams.fold([], |a, b| interleave(a, b))

The problem is that in Rust, `fold()` is not lazy, and that recursive
call to `interleave()` needs to be deferred otherwise it will cause a
stack overflow. `iter::Flatten` takes care of this by using the `Inner`
struct to defer the execution of interleaving two streams.

[1]: https://en.wikipedia.org/wiki/Hilbert%27s_paradox_of_the_Grand_Hotel