Extract `Thunk` in `iter::Flatten`
The `Inner` struct used for the `Flatten` type conflates a couple of
different concerns: implementing the `Iterator` trait, and differing the
internal call to interleave pairs of streams.

We can extract the latter into a struct named `Thunk`, which takes a
closure that represents something we'll need the value of later. The
first time `Thunk` is accessed, it executes the closure and stores its
result. Subsequent invocations return a reference to the forced value.