Trace the execution to display proofs of statements
We'd like to be able to display proofs of statements derived from a set
of rules. That is, a tree of the instances of rules used to derive the
statement, showing each rule with its conclusion, its premises, and
recursively the proofs of those premises.

To do this, we need to trace the execution to produce a tree of rules
and their conclusions. The struct use for this is `Proof`, which records
the name of a rule, the `State` when the rule was activated, the parent
proofs of the rule's premises, and the rule's conclusion. We build this
struct through the execution of `RuleSet::derive`.

Currently, rule execution works as follows. Say we have a rule with
conclusion C and premises A and B:

    C :- A, B.

To derive a target statement T in state S, we first unify T with the
rule's conclusion, producing a new state S'.

    let S' = S.unify(T, C)

Then we iterate over the premises. In general, deriving each premise
produces an infinite stream of states. We fold over the premises,
mapping an input stream of states to a new stream, letting the starting
state set be [S']. To produce a new stream for each premise, we derive
the premise for every state in the input stream, and flatten the result.

    [A, B].fold([S'], |states, premise| {
        let streams = state.map(|state| rules.derive(state, premise))
        flatten(streams)
    })

To build a tree of proofs, we need to keep the `Proof` of each premise,
and build a `Proof` for the conclusion with the premise proofs as its
parents. We replace the starting state S' with a pair of a `State` and
an empty list:

    let S' = (S.unify(T, C), [])

Then when folding over the premises, each call to `derive()` returns a
stream of `(State, Proof)` and we map this to a stream of `(State,
Vector<Proof>)` to collect the proofs of each premise together.

    let states = [A, B].fold([S'], |states, premise| {
        let streams = states.map(|(state, proofs)| {
            let proof_states = rules.derive(state, premise)
            proof_states.map(|(state, proof)| (state, proofs ++ [proof]))
        })
        flatten(streams)
    })

Finally we take this stream of `(State, Vector<Proof>)` and turn it into
a stream of `(State, Proof)` by wrapping those lists of proofs into a
single proof with the list as its parents.

    states.map(|(state, proofs)| {
        let proof = Proof.new(name, &state, proofs, &conclusion)
        (state, proof)
    })

This builds a tree of `Proof` structs that we can print to show the
trace of the program. We use `im::Vector<Rc<Parent>>` for the parents as
a single proof will end up being part of many different traces and this
avoids deep-cloning the whole tree through the above process.