Print proofs as trees
Here we take the existing `Proof::display` function and turn it into the
`Debug` impl for `Proof`. This function displays the proof as a simple
textual nested list.

For the `Display` impl, we implement a horizontal tree view that
resembles logical proof diagrams. For example, it can emit this output
to show the result of reversing a list of two items, `[b, c]`:

    -------------- rev-0   --------------------------- append-0       --------------------------- append-0
    rev nil eq nil         nil plus (c nil) eq (c nil)                nil plus (b nil) eq (b nil)
    -------------------------------------------------- rev-N      ----------------------------------- append-N
                  rev (c nil) eq (c nil)                          (c nil) plus (b nil) eq (c (b nil))
                  ----------------------------------------------------------------------------------- rev-N
                                            rev (b (c nil)) eq (c (b nil))