Implement a proper variable scope mechanism
Until now we'd been implementing scoping by appending a "scope number"
to the names of variables in the expressions for each rule. Here we
replace this with a `scope` field in `Variable` itself and a
`Expr::in_scope` function that generates a copy of an expression will
all variables given the correct scope.

We also make the `name` field into `Rc<String>` so that we don't copy
the `String` when creating scoped copies of a `Variable`.

It makes a minor difference to the derive benchmarks...

Before:
    test breadth_first_b ... bench:      63,105 ns/iter (+/- 10,810)
    test breadth_first_c ... bench:     969,124 ns/iter (+/- 183,567)
    test breadth_first_d ... bench:   9,268,414 ns/iter (+/- 1,610,858)
    test breadth_first_e ... bench: 103,304,238 ns/iter (+/- 27,115,662)

After:
    test breadth_first_b ... bench:      52,865 ns/iter (+/- 26,710)
    test breadth_first_c ... bench:     794,692 ns/iter (+/- 100,223)
    test breadth_first_d ... bench:   7,782,628 ns/iter (+/- 4,021,353)
    test breadth_first_e ... bench:  90,312,150 ns/iter (+/- 27,578,898)