Introduce a List type as a new core expression
A common thing I find myself using sequences for is to represent
dynamically-sized lists and sets by forming linked lists of sequences
with a marker terminator element.

For example, the list `[a, b, c]` might be written:

    (a (b (c nil)))

A record of key-value pairs `{ k = a, l = b, m = c }` might be written
as follows, where `rcd` is the null element and the general structure is
`(key = val tail)`

    (k = a (l = b (m = c rcd)))

There are several things we can do to improve this. First, it's hard to
read and write in explicit linked-list format -- it adds a lot of line
noise to expressions. Second, it has a negative impact on performance,
because `Sequence` is represented as an arbitrary-size `Vec` requiring
iteration, rather than a fixed two-element struct.

It is also expensive to tell two lists of different types apart, because
the thing that indicates the list's "type" is the terminator element
(`nil` or `rcd` above) is at the very end of a nested structure, so if
two lists differ only by their type it takes O(N) time to fail to unify
them, rather than O(1) as would be preferable.

Prolog has linked lists built in; a list can be destructured as `[Head |
Tail]`, and `[a, b, c]` is sugar for `[a | [b | [c | []]]]`. It does
suffer a usability problem in that lists often need wrapping in a "tag"
that needs to be explicitly propagated from `tag([Head | Tail])` to
`tag(Tail)` and vice versa, for example:

    length(list([]), 0).
    length(list([Head | Tail]), Len + 1) :- length(list(Tail), Len).

The struct and macros included here hint at a possible solution to this:
the tag is built into the `List` type itself, and the syntax `tag[a, b,
c]` desugars to propagate the tag to tail elements:

    tag[a, b, c] == tag[a | tag[b | tag[c | tag[]]]]

This means the above Prolog rules could be written as:

    length(list[], 0).
    length(list[Head | Tail], Len + 1) :- length(Tail, Len).

The `list()` wrapper does not need to be explicitly propagated in the
body of `length/2` because an expression like `list[a, b, c]` takes care
of that, and the assumption is that `Tail` will have the appropriate
tag automatically.