Benchmark lists against sequences
This benchmark compares the performance of appending two lists expressed
using sequences vs using tagged lists. For successful cases the
performance is roughly equal, but in failure cases tagged lists are much
faster and show constant-time performance compared to linear time for
sequences.

    test list_append_10             ... bench:     208,124 ns/iter (+/- 116,104)
    test list_append_100            ... bench:   1,666,186 ns/iter (+/- 270,859)
    test list_append_1_000          ... bench:  25,906,980 ns/iter (+/- 13,352,508)
    test list_append_fail_10        ... bench:       7,272 ns/iter (+/- 1,490)
    test list_append_fail_100       ... bench:       7,184 ns/iter (+/- 792)
    test list_append_fail_1_000     ... bench:       7,309 ns/iter (+/- 655)

    test sequence_append_10         ... bench:     155,517 ns/iter (+/- 106,927)
    test sequence_append_100        ... bench:   1,623,846 ns/iter (+/- 199,624)
    test sequence_append_1_000      ... bench:  25,964,999 ns/iter (+/- 12,926,772)
    test sequence_append_fail_10    ... bench:     133,285 ns/iter (+/- 13,439)
    test sequence_append_fail_100   ... bench:   1,542,044 ns/iter (+/- 233,610)
    test sequence_append_fail_1_000 ... bench:  23,676,075 ns/iter (+/- 9,018,499)

That's because sequences use their *last* element -- the "null" element
-- to indicate their type, so we need to unify the whole list before
discovering the types don't match.

This is not an especially common situation but we do need to be able to
efficiently express that an expression belongs to a certain syntactic
class, which sometimes comes town to comparing a type marker.