Don't call `eq` recursively in `State::unify`
The `unify_mut()` begins by checking `x == y`, which will do a recursive
equality check on the two expression structures. This will only pass if
the expressions are trivially identical, not if they're unifiable but
contain unbound variables. If it fails we then do another recursive scan
to unify them, which will also succeed if the expressions are identical.

This equality check is therefore wasted effort and we should do this
check at the leaves of expressions -- words and variables.