Extract `scope` from `Variable` struct
Executing logic programs requires a notion of "scope" for variables in
rules. Each time a rule is activated, we need to generate new instances
of all its variables so that they're not confused across different uses
of the same rule, or across different rules. We currently do this by
rewriting every expression to add a scope number to all variables, e.g.
the expression as written

    λ[$head | $tail] : $type

becomes:

    λ[$head/1 | $tail/1] : $type/1

and then we proceed with unifying it with the target. This causes a lot
of copying of expressions as we evaluation programs. Here we replace
this mechanism by storing scopes alongside expressions, rather than
within them. So, rather than rewrite the above expression, we bind it to
a scope using a tuple:

    (1, Rc(λ[$head | $tail] : $type))

The expression in this tuple does not need cloning because we use `Rc`
to pass references everywhere. When unifying, the scope is passed down
as we walk through the expression, and we store scopes alongside
variables and their values in `State`. An expression is always stored
with the scope it originated in, so we can resolve all its variables
later. This completely removes the need to copy expressions during
unification.

It has a slight negative impact on unification itself, because
`State::unify` has to do more work in unpacking and repacking these
scope structs instead of just unifying whatever opaque expressions it's
given.

Before:
    test nested_seq_with_late_unequal_word ... bench:         700 ns/iter (+/- 73)
    test nested_seqs                       ... bench:         691 ns/iter (+/- 72)
    test nested_seqs_with_indirect_vars    ... bench:       5,871 ns/iter (+/- 304)
    test seqs_with_vars                    ... bench:       2,477 ns/iter (+/- 766)
    test var                               ... bench:       1,499 ns/iter (+/- 571)
    test words                             ... bench:         502 ns/iter (+/- 23)

After:
    test nested_seq_with_late_unequal_word ... bench:         784 ns/iter (+/- 62)
    test nested_seqs                       ... bench:         787 ns/iter (+/- 41)
    test nested_seqs_with_indirect_vars    ... bench:       8,488 ns/iter (+/- 656)
    test seqs_with_vars                    ... bench:       2,920 ns/iter (+/- 199)
    test var                               ... bench:       1,477 ns/iter (+/- 989)
    test words                             ... bench:         470 ns/iter (+/- 45)

However, it has a positive impact on `RuleSet::derive`, which is where
we'd previously generate copies of expressions using `Expr::in_scope`:

Before:
    test breadth_first_b ... bench:      46,511 ns/iter (+/- 2,523)
    test breadth_first_c ... bench:     728,444 ns/iter (+/- 68,333)
    test breadth_first_d ... bench:   6,966,790 ns/iter (+/- 368,039)
    test breadth_first_e ... bench:  82,591,482 ns/iter (+/- 21,681,884)

After:
    test breadth_first_b ... bench:      37,223 ns/iter (+/- 2,364)
    test breadth_first_c ... bench:     618,421 ns/iter (+/- 31,897)
    test breadth_first_d ... bench:   5,907,782 ns/iter (+/- 436,099)
    test breadth_first_e ... bench:  70,131,580 ns/iter (+/- 14,707,447)

The numbers for the list vs sequence comparisons show a similar effect.

Before:
    test list_append_10             ... bench:     125,201 ns/iter (+/- 58,442)
    test list_append_100            ... bench:   1,435,443 ns/iter (+/- 134,465)
    test list_append_1_000          ... bench:  21,143,892 ns/iter (+/- 7,043,600)
    test list_append_fail_10        ... bench:       5,860 ns/iter (+/- 353)
    test list_append_fail_100       ... bench:       5,872 ns/iter (+/- 387)
    test list_append_fail_1_000     ... bench:       5,947 ns/iter (+/- 1,118)
    test sequence_append_10         ... bench:     128,400 ns/iter (+/- 89,048)
    test sequence_append_100        ... bench:   1,458,993 ns/iter (+/- 125,060)
    test sequence_append_1_000      ... bench:  20,714,333 ns/iter (+/- 1,080,107)
    test sequence_append_fail_10    ... bench:     117,195 ns/iter (+/- 7,115)
    test sequence_append_fail_100   ... bench:   1,432,182 ns/iter (+/- 135,481)
    test sequence_append_fail_1_000 ... bench:  19,552,299 ns/iter (+/- 10,754,603)

After:
    test list_append_10             ... bench:      89,829 ns/iter (+/- 7,399)
    test list_append_100            ... bench:   1,114,415 ns/iter (+/- 576,672)
    test list_append_1_000          ... bench:  16,701,954 ns/iter (+/- 1,294,892)
    test list_append_fail_10        ... bench:       2,450 ns/iter (+/- 1,672)
    test list_append_fail_100       ... bench:       2,439 ns/iter (+/- 126)
    test list_append_fail_1_000     ... bench:       2,437 ns/iter (+/- 199)
    test sequence_append_10         ... bench:      89,917 ns/iter (+/- 5,207)
    test sequence_append_100        ... bench:   1,146,162 ns/iter (+/- 721,283)
    test sequence_append_1_000      ... bench:  16,725,953 ns/iter (+/- 1,001,298)
    test sequence_append_fail_10    ... bench:      80,782 ns/iter (+/- 58,994)
    test sequence_append_fail_100   ... bench:   1,046,135 ns/iter (+/- 86,676)
    test sequence_append_fail_1_000 ... bench:  15,893,863 ns/iter (+/- 1,125,955)

Overall, despite making unification slower, it has a positive impact on
overall program execution and is worth keeping. It's possible there's a
way of implementing this in a way that's more opaque to `State::unify`
as well.