Move `Rc` wrapper into the `Expr` variants
The type `Rc<Expr>` appears all over the place, in the data model it's
the type of sequence elements and list head/tail, it's the type stored
in `State::values`, and it's the argument type to many functions. This
is because we need to have multiple references to arbitrary parts of
expressions, in their source rules and in states that bind them to
variables.

This results in a lot of boilerplate and sometimes double-pointers. For
example, `Variable` is a newtype around `Rc<String>`, so an value like

    Rc::new(Expr::Var(Variable(Rc::new(String::from("x")))))

is effectively just a double-pointer to the sting "x".

To remove these problems we move the `Rc` wrapper inside the variants of
the `Expr` enum, so that an `Expr` is just a pointer to one of the
expression types and cloning it is cheap.

This also makes the use of `Rc` more transparent since it doesn't appear
all over the code -- we just have `&Expr` being passed to functions so
changing the representation should be easier.