Extract `Invocation` to represent an instance of a rule being tried
This is motivated by my trying to refactor `Rule::match_target` into
smaller pieces, and finding that each stage of the application of a rule
requires access to that `Rule` and to the parent `RuleSet`, so that has
to be passed between functions.

The `Invocation` struct packages up the values that are needing during a
single application of a rule so that it's easier to implement this as a
few short methods instead of one big one.