Replace `Rc` with normal references in the `Expr` struct
This is an attempt to see what effect on performance it would have to
use normal references instead of `Rc` to pass expressions around. We
originally resorted to `Rc` because `RuleSet::derive` and `State::unify`
copied expressions a lot, for example to add scopes to variables or
replace them with sub-expressions. `Rc` let us share values between
`Expr` instances without deep-cloning them.

We've since removed all the copying these functions were doing;
`State::unify` does a single pass over the input expressions and only
reads from them, it doesn't copy anything. `Rc` is just an easy way to
have a lot of references to something without caring about owners vs
borrowers, but we're in a position to define those rules now.

Expressions are owned by the `Rule` that contains them, and therefore by
a `RuleSet`, so those structs contain owned values. We can remove the
`Rc` from `Expr` so that everything is stored inline, with the exception
of `Pair::{head,tail}`, which need boxing to avoid infinite sizing.
`State` stores the result of performing a unification and it can contain
references to things: `&Variable` and `&Expr`.

The code changes required here are fairly straightforward: we need to
add lifetimes to say that things survive as long as the `RuleSet` that
we're using to run the program. The most interesting addition is this
bit in `State::resolve_scoped`:

    let var: &'e Variable = unsafe { transmute(var) };

This is needed because `State<'e>::values` is a `HashMap` whose keys are
`(usize, &'e Variable)`, which does not implement `Borrow<(usize, &[any
lifetime] Variable)`. That means we can't pass `&(usize, &Variable)` to
`State::values::get()`, it must be `&(usize, &'e Variable)` unless we
invent a wrapper type to implement `Borrow` on.

Instead we use `transmute` to say that the argument `expr: &Expr` has
lifetime `'e` just so we can pass it to `State::values::get()`. I'm told
this is safe because we don't *store* the `expr` argument anywhere, we
only read from it during the execution of `resolve_scoped()`, during
which time the reference in `expr` cannot become invalid.

So does this improve performance? On the `unify` benchmarks, there's not
a substantial difference:

Before:
    test nested_seq_with_late_unequal_word ... bench:         673 ns/iter (+/- 62)
    test nested_seqs                       ... bench:         706 ns/iter (+/- 236)
    test nested_seqs_with_indirect_vars    ... bench:       3,237 ns/iter (+/- 213)
    test seqs_with_vars                    ... bench:       1,413 ns/iter (+/- 503)
    test var                               ... bench:         992 ns/iter (+/- 56)
    test words                             ... bench:         401 ns/iter (+/- 108)

After:
    test nested_seq_with_late_unequal_word ... bench:         542 ns/iter (+/- 34)
    test nested_seqs                       ... bench:         535 ns/iter (+/- 20)
    test nested_seqs_with_indirect_vars    ... bench:       3,045 ns/iter (+/- 633)
    test seqs_with_vars                    ... bench:       1,391 ns/iter (+/- 97)
    test var                               ... bench:         999 ns/iter (+/- 57)
    test words                             ... bench:         409 ns/iter (+/- 20)

The `derive` benchmarks show a very slight improvement:

Before:
    test breadth_first_b ... bench:      30,275 ns/iter (+/- 2,389)
    test breadth_first_c ... bench:     495,597 ns/iter (+/- 38,583)
    test breadth_first_d ... bench:   4,942,076 ns/iter (+/- 461,893)
    test breadth_first_e ... bench:  62,389,896 ns/iter (+/- 6,385,756)

After:
    test breadth_first_b ... bench:      27,772 ns/iter (+/- 1,259)
    test breadth_first_c ... bench:     450,970 ns/iter (+/- 22,926)
    test breadth_first_d ... bench:   4,339,027 ns/iter (+/- 575,848)
    test breadth_first_e ... bench:  54,424,276 ns/iter (+/- 6,516,002)

The most noticeable difference is in the benchmarks for sequences vs
lists, where a decent improvement is visible for larger problems.

Before:
    test list_append_10             ... bench:      66,890 ns/iter (+/- 2,466)
    test list_append_100            ... bench:     888,544 ns/iter (+/- 416,688)
    test list_append_1_000          ... bench:  14,257,277 ns/iter (+/- 4,158,357)
    test list_append_fail_10        ... bench:       2,280 ns/iter (+/- 1,760)
    test list_append_fail_100       ... bench:       2,288 ns/iter (+/- 392)
    test list_append_fail_1_000     ... bench:       2,264 ns/iter (+/- 148)
    test sequence_append_10         ... bench:      66,748 ns/iter (+/- 2,466)
    test sequence_append_100        ... bench:     889,451 ns/iter (+/- 75,580)
    test sequence_append_1_000      ... bench:  14,178,177 ns/iter (+/- 2,793,257)
    test sequence_append_fail_10    ... bench:      57,956 ns/iter (+/- 3,289)
    test sequence_append_fail_100   ... bench:     831,708 ns/iter (+/- 249,983)
    test sequence_append_fail_1_000 ... bench:  13,378,649 ns/iter (+/- 1,409,304)

After:
    test list_append_10             ... bench:      62,895 ns/iter (+/- 14,475)
    test list_append_100            ... bench:     794,265 ns/iter (+/- 83,485)
    test list_append_1_000          ... bench:  11,639,891 ns/iter (+/- 865,933)
    test list_append_fail_10        ... bench:       2,269 ns/iter (+/- 133)
    test list_append_fail_100       ... bench:       2,288 ns/iter (+/- 96)
    test list_append_fail_1_000     ... bench:       2,277 ns/iter (+/- 392)
    test sequence_append_10         ... bench:      62,734 ns/iter (+/- 2,684)
    test sequence_append_100        ... bench:     791,277 ns/iter (+/- 50,454)
    test sequence_append_1_000      ... bench:  11,708,667 ns/iter (+/- 2,783,442)
    test sequence_append_fail_10    ... bench:      59,912 ns/iter (+/- 14,257)
    test sequence_append_fail_100   ... bench:     785,205 ns/iter (+/- 119,798)
    test sequence_append_fail_1_000 ... bench:  11,056,155 ns/iter (+/- 2,820,057)

However I have a hunch that a better way to improve the performance on
these cases is to use a different search strategy that does a more
balanced tree traversal and doesn't take so long to resolve these
chained transitive rules.